Remake bootstrap as more general
Bootstrap now serves as a general deployment tool for the game, not just ecs worlds. This was done due to the fact that the InitializationGroup in the worlds is not suitable for a one-time call of the system OnCreate and the lack of a bootstrap functionality in Unity with the ability to sort bootstrap calls by branches.